feat: Premium logo ve gelişmiş setup installer
- Yeni büyük modern logo (512x512px) tasarlandı
- Hakkında modalında 200x200px container ile görüntü
- Premium uygulama ikonu (app_icon_new.png)
- Gelişmiş Inno Setup konfigürasyonu:
  * Setup sihirbazı ikonları ve görselleri
  * Hızlı başlatma alanı desteği
  * .opt dosya ilişkilendirme
  * Uninstall ikonu ve temiz kaldırma
  * Registry kayıtları
- Proje temizliği ve optimizasyonu
- Gereksiz dosyalar ve klasörler silindi
- Build ve run sorunları çözüldü

+++ ViewModels/MainViewModel.cs
@@ -1289,7 +1289,7 @@ namespace OptikFormApp.ViewModels
                 WrongDeductionFactor = _wrongDeductionFactor,
                 ThemeIndex = _themeIndex,
                 LayoutIndex = _layoutIndex,
-                FontSize = _fontSize
+                FontSize = (int)_fontSize
             });
         }
